Moved cave to activities folder
